Move amcl things into amcl_utils, fix lints
